refactor: convert gamePlayScreen function to gamePlayWidget Class
