+++ app/src/styles/widgets/gamePlayWidget.scss
@@ -1,5 +1,9 @@
 @use '../abstracts/variables' as *;
 
+.game-widget-container {
+  height: 100%;
+}
+
 .chat {
   display: flex;
   flex-direction: column;
