+++ app/src/ui/screens/game/gameScreen.ts
@@ -2,6 +2,7 @@ import { createDashboardLayout } from '../../layouts/dashboardLayout';
 import { createDashboardHeader } from '../dashboard/dashboardHeader';
 import createSidebar from '../dashboard/dashboardSideBar';
 import QuizWidget from '../widgets/quizWidget';
+import GamePlayWidget from '../widgets/gamePlayWidget';
 import { eventBus } from '../../../core/EventBus';
 
 export type GameScreenProps = {
@@ -40,18 +41,15 @@ export default function renderGameScreen({
   if (gameId === 'quiz') {
     const quiz = new QuizWidget(widgetContainer, gameId, onBack);
     quiz.start(skill, day);
-  } else {
-    widgetContainer.innerHTML = `<h2 style="color:white">Game ${gameId} is under construction!</h2><p style="color:white">The button simulates the successful completion of the game.</p>`;
-    const tempBackBtn = document.createElement('button');
-    tempBackBtn.textContent = '← Finish the game';
-    tempBackBtn.addEventListener('click', () => {
-      eventBus.emit('TASK_FINISHED', {
-        gameId: gameId,
-        outcome: 'correct',
-        userAnswer: `Simulated answer for ${gameId}`,
-      });
+  } else if (gameId === 'bugfix' || gameId === 'debug') {
+    const gameWidget = new GamePlayWidget(widgetContainer, {
+      day,
+      gameId: gameId as 'bugfix' | 'debug',
+      onBack,
     });
-    widgetContainer.append(tempBackBtn);
+    gameWidget.start();
+  } else {
+    widgetContainer.innerHTML = `<h2 style="color:white">Game ${gameId} is under construction!</h2>`;
   }
 
   layout.main.replaceChildren(widgetContainer);
